requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
requests>=2.32,<3.0
aiohttp>=3.10,<4.0
httpx>=0.28,<1.0
apscheduler>=3.10,<4.0
pillow==11.2.1
numpy==2.2.0
pyyaml==6.0.2
